chore: renamed template file
