+++ app/src/utils/csv-templates.ts
@@ -76,6 +76,11 @@ export const getSamplePeriodCSVTemplate = (): CSVEncodedTemplate => {
   return getCSVTemplate(['SAMPLE_SITE', 'TECHNIQUE_NAME', 'START_DATE', 'START_TIME', 'END_DATE', 'END_TIME']);
 };
 
+/**
+ * Get CSV template for Animals.
+ *
+ * @returns {CSVEncodedTemplate} Encoded CSV template
+ */
 export const getAnimalCSVTemplate = (): CSVEncodedTemplate => {
   return getCSVTemplate(['SPECIES', 'ALIAS', 'SEX', 'WLH_ID', 'DESCRIPTION']);
 };
